fix: payment vnpay

+++ NET1806_LittleJoyStore/NET1806_LittleJoy.Service/Services/VNPayService.cs
@@ -53,7 +53,7 @@ namespace NET1806_LittleJoy.Service.Services
             pay.AddRequestData("vnp_Version", _configuration["Vnpay:Version"]);
             pay.AddRequestData("vnp_Command", _configuration["Vnpay:Command"]);
             pay.AddRequestData("vnp_TmnCode", _configuration["Vnpay:TmnCode"]);
-            pay.AddRequestData("vnp_Amount", (price * 100).ToString());
+            pay.AddRequestData("vnp_Amount", ((double)price * 100).ToString());
             pay.AddRequestData("vnp_CreateDate", dateTime.ToString("yyyyMMddHHmmss"));
             pay.AddRequestData("vnp_CurrCode", _configuration["Vnpay:CurrCode"]);
             pay.AddRequestData("vnp_IpAddr", ipAddress);
